chore(sql): improve boundary checks for string and binary columns

+++ core/src/main/java/io/questdb/std/DirectLongList.java
@@ -159,7 +159,7 @@ public class DirectLongList implements Mutable, Closeable, Reopenable {
     }
 
     public long size() {
-        return (int) ((pos - address) >>> 3);
+        return (pos - address) >>> 3;
     }
 
     public void sortAsUnsigned() {
